+++ Pixen Application/Tools/Tool Palette/PXToolPaletteController.h
@@ -64,7 +64,7 @@
 - (void)keyDown:(NSEvent *)event fromCanvasController:(PXCanvasController *)cc;
 - (BOOL)keyWasDown:(unsigned int)mask;
 - (BOOL)isMask:(unsigned int)newMask upEventForModifierMask:(unsigned int)mask;
-- (BOOL)isMask:(unsigned int)newMask downEventForModifierMask:(unsigned int)mask;
+- (BOOL)isMask:(NSUInteger)newMask downEventForModifierMask:(unsigned int)mask;
 - (void)flagsChanged:(NSEvent *)theEvent;
 
 - (void)rightMouseDown;
